ml/pyproject.toml
[project]
name = "blockchain-bandits-ml"
version = "0.1.0"
description = "ML training + inference for Blockchain-Bandits"
requires-python = ">=3.11"
dependencies = [
  "numpy>=1.26.0",
  "pandas>=2.2.0",
  "polars>=1.9.0",
  "pyarrow>=17.0.0",
  "duckdb>=1.1.0",
  "scikit-learn>=1.5.0",
  "xgboost>=2.1.0",
  "shap>=0.46.0",
  "snorkel>=0.10.0",
  "networkx>=3.3",
  "wandb>=0.18.0",
  "modal>=0.64.0",
  "google-cloud-bigquery>=3.25.0",
  "python-dotenv>=1.0.1",
]

[project.optional-dependencies]
gpu = [
  # PyTorch + PyG installed separately on Colab; pinning here is brittle.
  # Run the bootstrap cells in notebooks/00_setup.py instead.
]
dev = [
  "pytest>=8.3.0",
  "ruff>=0.7.0",
  "ipykernel>=6.29.0",
  "jupyter>=1.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/blockchain_bandits_ml"]
